Move lit-html tests into render tests. Remove duplicates.
