+++ test/index.html
@@ -6,9 +6,8 @@
   </head>
   <body>
     <script type="module" src="./core_test.js"></script>
-    <script type="module" src="./lit-html_test.js"></script>
-    <script type="module" src="./lib/parts_test.js"></script>
     <script type="module" src="./lib/render_test.js"></script>
+    <script type="module" src="./lib/parts_test.js"></script>
     <script type="module" src="./lib/lit-extended_test.js"></script>
     <script type="module" src="./lib/modify-template_test.js"></script>
     <script type="module" src="./directives/repeat_test.js"></script>
